requirements.txt
## **ÉTAPE 3: Simplifier requirements.txt**

fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
pillow==12.1.0
stripe==14.1.0
python-dotenv==1.2.1
aiofiles==23.2.1
rembg==2.0.72
numpy==2.3.0
onnxruntime==1.18.0
torch==2.4.0
torchvision==0.19.0
requests==2.32.5